+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support for CAPI kubeconfig structure
+
 ## [4.0.0] - 2020-12-11
 
 ### Changed